-Moved BFM Lanczos into its own namespace to avoid conflicts with Grid -Imported changes from Luchang's Lanczos -Various other BFM lanczos improvements -Removed old A2A code -Performance improvements for spincolorflavormatrix -Additional matrix methods used by new A2A code -Added FGrid

+++ cps_pp/src/util/lattice/fix_gauge/fix_gauge.C
@@ -1010,7 +1010,7 @@ IFloat FixHPlane::delta(int recurse)
 	Matrix& A = findA();   // reference to A_buf
 	tmp_m.Dagger(A);
 	A -= tmp_m;
-	A -= A.Tr()/COLORS;
+	A -= A.Tr()/(double)COLORS;
 	for(int i=0; i<COLORS*COLORS; i++)
 	  dlt += norm(A[i]);
       }
